Removed JS support via ESLint
- Also convert the `markdown` module to TS
- Add `allowImportingTsExtensions`, safe since we use bundler mode
- Bump deps
