+++ package.json
@@ -16,16 +16,16 @@
     ]
   },
   "devDependencies": {
-    "@eslint/js": "^9.31.0",
     "@jimmyverburgt/svelte-input-otp": "^0.0.3",
     "@paypal/paypal-js": "^8.2.0",
-    "@stylistic/eslint-plugin": "^5.1.0",
+    "@stylistic/eslint-plugin": "^5.2.0",
     "@sveltejs/adapter-node": "^5.2.13",
     "@sveltejs/kit": "^2.24.0",
     "@sveltejs/vite-plugin-svelte": "^5.1.0",
     "@types/node": "^24.0.14",
     "@types/postcss-import": "^14.0.3",
     "@types/semver": "^7.7.0",
+    "@types/unist": "^3.0.3",
     "@typescript-eslint/eslint-plugin": "^8.37.0",
     "@typescript-eslint/parser": "^8.37.0",
     "eslint": "^9.31.0",
@@ -38,7 +38,7 @@
     "postcss": "^8.5.6",
     "postcss-import": "^16.1.1",
     "semver": "^7.7.2",
-    "svelte": "^5.36.1",
+    "svelte": "^5.36.2",
     "svelte-eslint-parser": "^1.2.0",
     "svelte-preprocess": "^6.0.3",
     "svelte-turnstile": "^0.11.0",
